+++ components/ILIAS/Container/Block/class.ilCustomBlock.php
