+++ components/ILIAS/Container/resources/ilblockcallback.js
